+++ WalletWasabi.Fluent/Controls/Payment/BtcAndPayjoinPaymentControl.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
-             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.DestinationEntry.Components"
+             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.Payment.Components"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="WalletWasabi.Fluent.Controls.DestinationEntry.BtcAndPayjoinPaymentControl"
+             x:Class="WalletWasabi.Fluent.Controls.Payment.BtcAndPayjoinPaymentControl"
              x:Name="UserControl">
 
   <UserControl.Styles>
