+++ WalletWasabi.Fluent/Controls/Payment/BtcPaymentControl.axaml
@@ -2,9 +2,9 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.DestinationEntry.Components"
+             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.Payment.Components"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="WalletWasabi.Fluent.Controls.DestinationEntry.BtcPaymentControl"
+             x:Class="WalletWasabi.Fluent.Controls.Payment.BtcPaymentControl"
              x:Name="UserControl">
 
   <UserControl.Styles>
