+++ WalletWasabi.Fluent/Controls/Payment/Components/AddressEntryControl.axaml
@@ -4,9 +4,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:b="clr-namespace:WalletWasabi.Fluent.Behaviors"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.DestinationEntry.Components"
+             xmlns:components="clr-namespace:WalletWasabi.Fluent.Controls.Payment.Components"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="WalletWasabi.Fluent.Controls.DestinationEntry.Components.AddressEntryControl"
+             x:Class="WalletWasabi.Fluent.Controls.Payment.Components.AddressEntryControl"
              x:Name="UserControl">
   <DockPanel>
     <DockPanel>
