+++ WalletWasabi.Fluent/Controls/Payment/Components/QrCodeButton.axaml.cs
@@ -3,9 +3,9 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 using ReactiveUI;
-using WalletWasabi.Fluent.Controls.DestinationEntry.ViewModels;
+using WalletWasabi.Fluent.Controls.Payment.ViewModels;
 
-namespace WalletWasabi.Fluent.Controls.DestinationEntry.Components;
+namespace WalletWasabi.Fluent.Controls.Payment.Components;
 
 public class QrCodeButton : UserControl
 {
