+++ WalletWasabi.Fluent/Controls/Payment/ViewModels/AmountCurrencyDirectionController.cs
@@ -3,20 +3,21 @@ using System.Reactive.Linq;
 using Avalonia.Controls.Mixins;
 using ReactiveUI;
 
-namespace WalletWasabi.Fluent.Controls.DestinationEntry.ViewModels;
+namespace WalletWasabi.Fluent.Controls.Payment.ViewModels;
 
 public class AmountCurrencyDirectionController : ReactiveObject, IDisposable
 {
-	private bool _isReversed;
 	private readonly CompositeDisposable _disposables = new();
+	private bool _isReversed;
 
 	public AmountCurrencyDirectionController(bool initialValue, Action<bool> setIsReversed)
 	{
 		IsReversed = initialValue;
 		DisposableMixin.DisposeWith(
-				this.WhenAnyValue(x => x.IsReversed)
-					.Do(setIsReversed)
-					.Subscribe(), _disposables);
+			this.WhenAnyValue(x => x.IsReversed)
+				.Do(setIsReversed)
+				.Subscribe(),
+			_disposables);
 	}
 
 	public bool IsReversed
