+++ WalletWasabi.Fluent/Controls/Payment/ViewModels/ScanQrViewModel.cs
@@ -5,7 +5,7 @@ using ReactiveUI;
 using WalletWasabi.Fluent.ViewModels.Dialogs;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 
-namespace WalletWasabi.Fluent.Controls.DestinationEntry.ViewModels;
+namespace WalletWasabi.Fluent.Controls.Payment.ViewModels;
 
 public class ScanQrViewModel
 {
