+++ WalletWasabi.Fluent/Helpers/Clipboard.cs
@@ -1,0 +1,19 @@
+using System.Reactive.Linq;
+using ReactiveUI;
+
+namespace WalletWasabi.Fluent.Helpers;
+
+public class Clipboard
+{
+	public Clipboard()
+	{
+		ContentChanged = Observable
+			.Timer(TimeSpan.FromMilliseconds(200), RxApp.MainThreadScheduler)
+			.Repeat()
+			.SelectMany(_ => ApplicationUtils.GetClipboardTextAsync())
+			.Select(x => x.Trim())
+			.DistinctUntilChanged();
+	}
+
+	public IObservable<string> ContentChanged { get; }
+}
