+++ WalletWasabi.Fluent/Helpers/IAddressParser.cs
@@ -1,4 +1,6 @@
-namespace WalletWasabi.Fluent.Controls.DestinationEntry.ViewModels;
+using WalletWasabi.Fluent.Models;
+
+namespace WalletWasabi.Fluent.Helpers;
 
 public interface IAddressParser
 {
