+++ WalletWasabi.Tests/UnitTests/Payment/AddressViewModelTests.cs
